+++ src/index.ts
@@ -1,8 +1,6 @@
 /**
  * Express middlewares
- * @module
  */
-
 import basicAuth from './middlewares/basicAuth';
 import cognitoAuth from './middlewares/cognitoAuth';
 import rateLimit from './middlewares/rateLimit';