+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+
 ### Added
 
 ### Changed
@@ -16,12 +18,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v0.1.1 - 2019-04-01
+
+### Changed
+
+- rateLimitミドルウェアのRedisキー名を変更(ディリミタをドットからコロンに変更)
+- update packages
 
 ## v0.1.0 - 2018-02-04
+
 ### Added
+
 - Cognito認証ミドルウェアを追加。
 
 ## v0.0.0 - 2017-11-28
+
 ### Added
+
 - ベーシック認証ミドルウェアを追加。
 - レート制限ミドルウェアを追加。